requirements.txt
discord.py>=2.3.2
python-dotenv
numexpr
requests
gtts
PyNaCl
ffmpeg
yt_dlp